requirements.txt
asgi-redis==0.10.0
channels==0.10.3
Django>=1.8,<1.9
django-extensions==1.6.1
django-flat-theme==1.1.3
django-redis-cache==1.6.5
djangorestframework==3.3.3
gevent==1.1.0
gunicorn==19.4.5
huey==1.1.1
mysqlclient==1.3.6
pylast==1.4.2
python-dateutil==2.5.2
redis==2.10.5
